+++ armonaut/project/models.py
@@ -22,7 +22,7 @@ from armonaut.db import Model
 class ProjectRoleType(enum.Enum):
     OWNER = 'owner'  # read, commit, admin
     COLLABORATOR = 'collaborator'  # read, commit
-    READ_ONLY = 'read_only'
+    READ_ONLY = 'read_only'  # read
 
 
 class ProjectRole(Model):